Refactor: Simplify semantic and morphology modules
- Flattened `src/morphology/matcher.rs` into `src/morphology/mod.rs` to reduce file count.
- Merged `src/semantic/types.rs` into `src/semantic/model.rs` to unify the semantic model and prevent circular dependencies.
- Renamed `src/semantic/conversion.rs` to `src/semantic/statements.rs` to better reflect its domain responsibility (Statement Analysis).
- Updated documentation and tests to reflect these changes.
- Fixed documentation tests in `model.rs` to use public re-exports.
