+++ src/semantic/expressions.rs
@@ -17,9 +17,8 @@ use crate::errors::GlossaError;
 use crate::morphology::{self, DisambiguationContext, analyze_article, disambiguate, resolve_best};
 use crate::semantic::assembler::Assembler;
 use crate::semantic::assembler::Literal;
-use crate::semantic::model::{AnalyzedExpr, AnalyzedExprKind};
+use crate::semantic::model::{AnalyzedExpr, AnalyzedExprKind, GlossaType};
 use crate::semantic::resolver::Scope;
-use crate::semantic::types::GlossaType;
 
 /// Analyze an argument expression (could be literal, variable, or nested call)
 ///
