+++ src/semantic/statements_tests.rs
@@ -1,4 +1,4 @@
-use super::conversion::extract_value;
+use super::statements::extract_value;
 use crate::ast::Expr;
 use crate::morphology::Case;
 use crate::morphology::lexicon::BinaryOp;
